+++ src/client/client.rs
@@ -540,10 +540,8 @@ impl<C: ClientConnection> Client<C> {
   ///
   /// # Arguments
   ///
-  /// * `name` - the record name to create
-  /// * `zone` - the zone name (must match an SOA) to update
-  /// * `class` - the DNS class
-  /// * `record_data` - the RData to associate to the new record
+  /// * `record` - the name of the record to create
+  /// * `zone_origin` - the zone name (must match an SOA) to update
   /// * `signer` - the signer with private key to use to sign the request
   ///
   /// The update must go to a zone authority (i.e. the server used in the ClientConnection)
@@ -587,6 +585,89 @@ impl<C: ClientConnection> Client<C> {
     self.send_message(&message)
   }
 
+  /// Appends a record to an existing rrset, optionally require the rrset to exist.
+  ///
+  /// [RFC 2136](https://tools.ietf.org/html/rfc2136), DNS Update, April 1997
+  ///
+  /// ```text
+  /// 2.4.1 - RRset Exists (Value Independent)
+  ///
+  ///   At least one RR with a specified NAME and TYPE (in the zone and class
+  ///   specified in the Zone Section) must exist.
+  ///
+  ///   For this prerequisite, a requestor adds to the section a single RR
+  ///   whose NAME and TYPE are equal to that of the zone RRset whose
+  ///   existence is required.  RDLENGTH is zero and RDATA is therefore
+  ///   empty.  CLASS must be specified as ANY to differentiate this
+  ///   condition from that of an actual RR whose RDLENGTH is naturally zero
+  ///   (0) (e.g., NULL).  TTL is specified as zero (0).
+  ///
+  /// 2.5.1 - Add To An RRset
+  ///
+  ///    RRs are added to the Update Section whose NAME, TYPE, TTL, RDLENGTH
+  ///    and RDATA are those being added, and CLASS is the same as the zone
+  ///    class.  Any duplicate RRs will be silently ignored by the primary
+  ///    master.
+  /// ```
+  ///
+  /// # Arguments
+  ///
+  /// * `record` - the record to append to an RRSet
+  /// * `zone` - the zone name (must match an SOA) to update
+  /// * `must_exist` - if true, the request will fail if the record does not exist
+  /// * `signer` - the signer with private key to use to sign the request
+  ///
+  /// The update must go to a zone authority (i.e. the server used in the ClientConnection). If
+  /// the rrset does not exist and must_exist is false, then the RRSet will be created.
+  pub fn append(&self,
+                record: Record,
+                zone_origin: domain::Name,
+                must_exist: bool,
+                signer: &Signer) -> ClientResult<Message> {
+    assert!(zone_origin.zone_of(record.get_name()));
+
+    // for updates, the query section is used for the zone
+    let mut zone: Query = Query::new();
+    zone.name(zone_origin).query_class(record.get_dns_class()).query_type(record.get_rr_type());
+
+    // build the message
+    let mut message: Message = Message::new();
+    message.id(self.next_id()).message_type(MessageType::Query).op_code(OpCode::Update).recursion_desired(false);
+    message.add_zone(zone);
+
+    if must_exist {
+      let mut prerequisite = Record::with(record.get_name().clone(), record.get_rr_type(), 0);
+      prerequisite.dns_class(DNSClass::ANY);
+      message.add_pre_requisite(prerequisite);
+    }
+
+    message.add_update(record);
+
+    // Extended dns
+    let mut edns: Edns = Edns::new();
+
+    // if secure {
+    //   edns.set_dnssec_ok(true);
+    //   message.authentic_data(true);
+    //   message.checking_disabled(false);
+    // }
+
+    edns.set_max_payload(1500);
+    edns.set_version(0);
+
+    message.set_edns(edns);
+
+    // after all other updates to the message, sign it.
+    message.sign(signer, UTC::now().timestamp() as u32);
+
+    self.send_message(&message)
+  }
+
+  /// Sends a message to the server for which this client was defined
+  ///
+  /// # Arguments
+  ///
+  /// * `message` - the message to deliver
   fn send_message(&self, message: &Message) -> ClientResult<Message> {
     // get the message bytes and send the query
     let mut buffer: Vec<u8> = Vec::with_capacity(512);
@@ -606,6 +687,7 @@ impl<C: ClientConnection> Client<C> {
     Ok(response)
   }
 
+  /// increments the next_id for use in messages
   fn next_id(&self) -> u16 {
     let id = self.next_id.get();
     self.next_id.set(id + 1);
@@ -857,13 +939,12 @@ mod test {
   //   assert_eq!(response.get_response_code(), ResponseCode::NXDomain);
   // }
 
-  #[test]
-  fn test_create() {
+  #[cfg(test)]
+  fn create_sig0_ready_client<'a>(catalog: &'a mut Catalog) -> (Client<TestClientConnection<'a>>, Signer, domain::Name) {
     use ::rr::rdata::DNSKEY;
 
     let mut authority = create_example();
     authority.set_allow_update(true);
-    let mut catalog = Catalog::new();
     let origin = authority.get_origin().clone();
 
     let mut pkey = PKey::new();
@@ -883,20 +964,89 @@ mod test {
     authority.upsert(auth_key, 0);
 
     catalog.upsert(authority.get_origin().clone(), authority);
-    let client = Client::new(TestClientConnection::new(&catalog));
+    let client = Client::new(TestClientConnection::new(catalog));
 
-    // add a record
+    (client, signer, origin)
+  }
+
+  #[test]
+  fn test_create() {
+    let mut catalog = Catalog::new();
+    let (client, signer, origin) = create_sig0_ready_client(&mut catalog);
+
+    // create a record
     let mut record = Record::with(domain::Name::with_labels(vec!["new".to_string(), "example".to_string(), "com".to_string()]),
                                   RecordType::A,
                                   Duration::minutes(5).num_seconds() as u32);
     record.rdata(RData::A(Ipv4Addr::new(100,10,100,10)));
 
 
-    let result = client.create(record.clone(), origin, &signer).expect("create failed");
+    let result = client.create(record.clone(), origin.clone(), &signer).expect("create failed");
     assert_eq!(result.get_response_code(), ResponseCode::NoError);
     let result = client.query(record.get_name(), record.get_dns_class(), record.get_rr_type()).expect("query failed");
     assert_eq!(result.get_response_code(), ResponseCode::NoError);
     assert_eq!(result.get_answers().len(), 1);
     assert_eq!(result.get_answers()[0], record);
+
+    // trying to create again should error
+    // TODO: it would be cool to make this
+    let result = client.create(record.clone(), origin.clone(), &signer).expect("create failed");
+    assert_eq!(result.get_response_code(), ResponseCode::YXRRSet);
+
+    // will fail if already set and not the same value.
+    let mut record = record.clone();
+    record.rdata(RData::A(Ipv4Addr::new(101,11,101,11)));
+
+    let result = client.create(record.clone(), origin.clone(), &signer).expect("create failed");
+    assert_eq!(result.get_response_code(), ResponseCode::YXRRSet);
+
+  }
+
+  #[test]
+  fn test_append() {
+    let mut catalog = Catalog::new();
+    let (client, signer, origin) = create_sig0_ready_client(&mut catalog);
+
+    // append a record
+    let mut record = Record::with(domain::Name::with_labels(vec!["new".to_string(), "example".to_string(), "com".to_string()]),
+                                  RecordType::A,
+                                  Duration::minutes(5).num_seconds() as u32);
+    record.rdata(RData::A(Ipv4Addr::new(100,10,100,10)));
+
+    // first check the must_exist option
+    let result = client.append(record.clone(), origin.clone(), true, &signer).expect("append failed");
+    assert_eq!(result.get_response_code(), ResponseCode::NXRRSet);
+
+    // next append to a non-existent RRset
+    let result = client.append(record.clone(), origin.clone(), false, &signer).expect("append failed");
+    assert_eq!(result.get_response_code(), ResponseCode::NoError);
+
+    // verify record contents
+    let result = client.query(record.get_name(), record.get_dns_class(), record.get_rr_type()).expect("query failed");
+    assert_eq!(result.get_response_code(), ResponseCode::NoError);
+    assert_eq!(result.get_answers().len(), 1);
+    assert_eq!(result.get_answers()[0], record);
+
+    // will fail if already set and not the same value.
+    let mut record = record.clone();
+    record.rdata(RData::A(Ipv4Addr::new(101,11,101,11)));
+
+    let result = client.append(record.clone(), origin.clone(), true, &signer).expect("create failed");
+    assert_eq!(result.get_response_code(), ResponseCode::NoError);
+
+    let result = client.query(record.get_name(), record.get_dns_class(), record.get_rr_type()).expect("query failed");
+    assert_eq!(result.get_response_code(), ResponseCode::NoError);
+    assert_eq!(result.get_answers().len(), 2);
+
+    assert!(result.get_answers().iter().any(|rr| if let &RData::A(ref ip) = rr.get_rdata() { *ip ==  Ipv4Addr::new(100,10,100,10) } else { false }));
+    assert!(result.get_answers().iter().any(|rr| if let &RData::A(ref ip) = rr.get_rdata() { *ip ==  Ipv4Addr::new(101,11,101,11) } else { false }));
+
+    // show that appending the same thing again is ok, but doesn't add any records
+    let result = client.append(record.clone(), origin.clone(), true, &signer).expect("create failed");
+    assert_eq!(result.get_response_code(), ResponseCode::NoError);
+
+    let result = client.query(record.get_name(), record.get_dns_class(), record.get_rr_type()).expect("query failed");
+    assert_eq!(result.get_response_code(), ResponseCode::NoError);
+    assert_eq!(result.get_answers().len(), 2);
   }
 }